Some cleanup.
